Make player damageable

+++ Assets/Scripts/Interfaces/IDamageable.cs
@@ -2,6 +2,6 @@ namespace Interfaces
 {
     public interface IDamageable
     {
-        void Damage(float damage);
+        void Damage(int damage);
     }
 }
